Redis connection fix (#7)
Laravel doesnt have ability to use different redis-clients for different connections. It can be problem if you want to use sentinel-connection for app and phpredis for metrics.

So, in the PR i changed redis config for Prometheus. Now, you can specify redis-client and all credentials right in event-tracker.php config

---------

Co-authored-by: umbr <aleksander.nau@cadolabs.io>